refactor: move template funcions to a new go file
issue: #40
